+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/BoxTestGenerated.java
@@ -45,6 +45,22 @@ public class BoxTestGenerated extends AbstractBoxTest {
   }
 
   @Nested
+  @TestMetadata("compiler-tests/src/test/data/box/contributesgraphextension")
+  @TestDataPath("$PROJECT_ROOT")
+  public class Contributesgraphextension {
+    @Test
+    public void testAllFilesPresentInContributesgraphextension() {
+      KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("compiler-tests/src/test/data/box/contributesgraphextension"), Pattern.compile("^(.+)\\.kt$"), null, TargetBackend.JVM_IR, true);
+    }
+
+    @Test
+    @TestMetadata("ContributingMultipleGraphsToSameParent.kt")
+    public void testContributingMultipleGraphsToSameParent() {
+      runTest("compiler-tests/src/test/data/box/contributesgraphextension/ContributingMultipleGraphsToSameParent.kt");
+    }
+  }
+
+  @Nested
   @TestMetadata("compiler-tests/src/test/data/box/cycles")
   @TestDataPath("$PROJECT_ROOT")
   public class Cycles {